용어,카테고리 테이블 관련 수정

+++ vita/src/main/java/com/ict/vita/repository/terms/TermsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 //[용어 Repository]
-public interface TermsRepository extends JpaRepository<TermsEntity, Integer>{
+public interface TermsRepository extends JpaRepository<TermsEntity, Long>{
 
 }
